Incomplete creation of FileProcessor object
